+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 *   (bc) Changed the structure and keys in the dependencies map, to also allow SCSS files to be written into the dependencies map.
 *   (feature) Add option to disable the legacy build.
 *   (improvement) Bump dependencies.
+*   (feature) Add option to force-enable TypeScript support.
 
 
 1.2.12